Improve user entity ownership code
