+++ src/Entity/Project/RewardClaim.php
@@ -3,6 +3,7 @@
 namespace App\Entity\Project;
 
 use App\Entity\Interface\UserOwnedInterface;
+use App\Entity\Trait\UserOwnedTrait;
 use App\Entity\User\User;
 use App\Mapping\Provider\EntityMapProvider;
 use App\Repository\Project\RewardClaimRepository;
@@ -13,6 +14,8 @@ use Doctrine\ORM\Mapping as ORM;
 #[ORM\Entity(repositoryClass: RewardClaimRepository::class)]
 class RewardClaim implements UserOwnedInterface
 {
+    use UserOwnedTrait;
+
     #[ORM\Id]
     #[ORM\GeneratedValue]
     #[ORM\Column]
@@ -39,23 +42,6 @@ class RewardClaim implements UserOwnedInterface
     public function setReward(?Reward $reward): static
     {
         $this->reward = $reward;
-
-        return $this;
-    }
-
-    public function getOwner(): ?User
-    {
-        return $this->owner;
-    }
-
-    public function isOwnedBy(User $user): bool
-    {
-        return $this->owner->getId() === $user->getId();
-    }
-
-    public function setOwner(?User $owner): static
-    {
-        $this->owner = $owner;
 
         return $this;
     }
