+++ src/Entity/Trait/UserOwnedTrait.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Entity\Trait;
+
+use App\Entity\User\User;
+
+trait UserOwnedTrait
+{
+    public function getOwner(): ?User
+    {
+        return $this->owner;
+    }
+
+    public function setOwner(?User $owner): static
+    {
+        $this->owner = $owner;
+
+        return $this;
+    }
+
+    public function isOwnedBy(User $user): bool
+    {
+        return $this->owner->getId() === $user->getId();
+    }
+}
